+++ Source/Scene/BingMapsImageryProvider.js
@@ -369,7 +369,7 @@ define([
         }
         //>>includeEnd('debug');
 
-        return this._tilingScheme.getExtent();
+        return this._tilingScheme.extent;
     };
 
     /**